+++ packages/web3-types/CHANGELOG.md
@@ -175,3 +175,4 @@ Documentation:
 ### Added
 
 -   Interface `MetaMaskProvider` added and is part of `SupportedProviders` (#6534)
+-   `gasPrice` was added to `Transaction1559UnsignedAPI` type. (#6539)