+++ packages/web3-types/CHANGELOG.md
@@ -173,4 +173,6 @@ Documentation:
 ## [Unreleased]
 
 ### Added
+
+-   Interface `MetaMaskProvider` added and is part of `SupportedProviders` (#6534)
 -   `gasPrice` was added to `Transaction1559UnsignedAPI` type. (#6539)